Changed the oppianapp management code so that it only creates the photo inline type if not already created.
Added creation of a small oppian o logo photo as well as a very basic initial blog post.

Needs some tidying and should add scanning of the initial images into photos.
Also should fill in the blog picture by finding the relevant photo rather than relying on it having the ID of 1.  This may require rearranging the installed apps to ensure that they are created in the right order.