Put the site name setting into the settings.py file, so that it can be overridden in settings_local.py when not using port 80 (and for when the hostname doesn't resolve properly e.g. on Mac)
Abstracted out the saving of the blog post from the the PostAdmin class - so that it can be used during syncdb to create a twitter status change.

** oppian app management
Added an oppianapp models.py, so that it can be used as a signal in its management code.  This way we can ensure that we are the last app and so the last thing to be initialised.
Made the oppianapp management code only print out information if the verbosity warrants it.
Made the basic.media initialisation scan the images folder and create the relevant photo entry in the DB, so that it can be used for blog postings.
Made the blog post creation happen within the oppianapp initialisation, so that we can find the correct photo object to put in it.