Created a new pair of template tags - set_vars_from_json and set_vars_from_json_block (housed in the set_vars_from_json module). These tags are used to read in a json obect from a template into the context.
These are used to simplify the sub menu structure - the list of menu items is created from a json object and then the template code to use them is included afterwards.
This template can also be shared between pages.
The url is used to select the correct tab in the list.
(I added the context processor django.core.context_processors.request to enable this).

I've also separated out the recommended, partners and map sections into separate pages, so that it matches the style of the people page in the about view.
We can always mmove these back easily is required.